Refactor blueprint's file structure
