+++ pilotecmwf_pp_starter/config/service_settings.py
@@ -3,8 +3,25 @@ from mch_python_commons.config.base_settings import BaseServiceSettings
 from pydantic import BaseModel
 
 
+class S3Bucket(BaseModel):
+    endpoint_url: str
+    name: str
+
+
+class S3Buckets(BaseModel):
+    input: S3Bucket
+    output: S3Bucket
+
+
+class TimeSettings(BaseModel):
+    tincr: int
+    tstart: int
+
+
 class AppSettings(BaseModel):
     app_name: str
+    s3_buckets: S3Buckets
+    time_settings: TimeSettings
 
 
 class ServiceSettings(BaseServiceSettings):
