--- pilotecmwf_pp_starter/domain/greeting.py
@@ -1,5 +1,0 @@
-from pydantic import BaseModel  # pylint: disable=no-name-in-module
-
-
-class Greeting(BaseModel):
-    message: str
